update absyn and add unitest
update absyn by fixing warning and add Quantified Term

Add unitest for manually building Hurricane model

+++ src/absyn/ArrayExpr.cpp
@@ -21,7 +21,7 @@ void ArrayExpr::print(FILE* file, int indent) {
 	fprintf(file, "%*s(ArrayExpr:\n", indent, "");
 	fprintf(file, "%*s:dim %d\n", indent+2, "", dim);
 	fprintf(file, "%*s(args:\n", indent + 2, "");
-	for (int i = 0; i < args.size(); i++)
+	for (size_t i = 0; i < args.size(); i++)
 		if (args[i] != NULL)
 			args[i]->print(file, indent + 4);
 	fprintf(file, "%*s):\n", indent + 2, "");
