+++ src/absyn/BlogProgram.cpp
@@ -14,12 +14,12 @@ BlogProgram::BlogProgram(int l, int c)
 }
 
 BlogProgram::~BlogProgram() {
-	for (int i = 0; i < args.size(); i++)
+	for (size_t i = 0; i < args.size(); i++)
 		if (args[i] != NULL)
 			delete args[i];
 }
 
-int BlogProgram::size() {
+size_t BlogProgram::size() {
 	return args.size();
 }
 
@@ -35,7 +35,7 @@ Decl* BlogProgram::get(int k) {
 void BlogProgram::print(FILE* file, int indent) {
 	fprintf(file, "%*s(BlogProgram:\n", indent, "");
 	fprintf(file, "%*s(args:\n", indent + 2, "");
-	for (int i = 0; i < args.size(); i++)
+	for (size_t i = 0; i < args.size(); i++)
 		if (args[i] != NULL)
 			args[i]->print(file, indent + 4);
 	fprintf(file, "%*s)\n", indent + 2, "");
