+++ src/absyn/BlogProgram.h
@@ -15,8 +15,31 @@
 #include <vector>
 using std::vector;
 
-#include "Absyn.h"
-#include "Decl.h"
+#include "ArrayExpr.h"
+#include "BoolLiteral.h"
+#include "DistinctDecl.h"
+#include "DistrExpr.h"
+#include "DoubleLiteral.h"
+#include "FuncApp.h"
+#include "FuncDecl.h"
+#include "IfExpr.h"
+#include "IntLiteral.h"
+#include "MapExpr.h"
+#include "NullLiteral.h"
+#include "NumStDecl.h"
+#include "NumStRef.h"
+#include "ObsDecl.h"
+#include "OpExpr.h"
+#include "OriginDecl.h"
+#include "QuantExpr.h"
+#include "QueryDecl.h"
+#include "SetExpr.h"
+#include "StringLiteral.h"
+#include "Symbol.h"
+#include "TimeStampLiteral.h"
+#include "TypDecl.h"
+#include "VarDecl.h"
+#include "VarRef.h"
 
 namespace swift {
 
@@ -28,7 +51,7 @@ public:
 	BlogProgram(int l, int r);
 	virtual ~BlogProgram();
 
-	int size();
+	size_t size();
 	void add(Decl* decl);
 	Decl* get(int k);
