+++ src/absyn/DistinctDecl.cpp
@@ -21,7 +21,7 @@ void DistinctDecl::add(Symbol v, int r) {
 	rep.push_back(r);
 }
 
-int DistinctDecl::size() {
+size_t DistinctDecl::size() {
 	return var.size();
 }
 
@@ -42,7 +42,7 @@ void DistinctDecl::print(FILE* file, int indent) {
 	fprintf(file, "%*s(DistinctDecl:\n", indent, "");
 	fprintf(file, "%*s:type %s\n", indent + 2 , "", typ.getValue().c_str());
 	fprintf(file, "%*s:var (:", indent + 2, "");
-	for (int i = 0; i < var.size(); i++)
+	for (size_t i = 0; i < var.size(); i++)
 		fprintf(file, " %s", var[i].getValue().c_str());
 	fprintf(file, " )\n");
 	fprintf(file, "%*s)\n", indent, "");
