+++ src/absyn/FuncApp.cpp
@@ -25,7 +25,7 @@ void FuncApp::print(FILE* file, int indent) {
 	fprintf(file, "%*s(FuncApp:\n", indent, "");
 	fprintf(file, "%*s:func %s\n", indent + 2, "", func.getValue().c_str());
 	fprintf(file, "%*s(args:\n", indent + 2, "");
-	for (int i = 0; i < args.size(); i++)
+	for (size_t i = 0; i < args.size(); i++)
 	if (args[i] != NULL)
 		args[i]->print(file, indent + 4);
 	fprintf(file, "%*s)\n", indent + 2, "");
